MEAN App
Remove of EventEmitter and output event from post-create component, as well as
Input event from post-list component. Likewise, chain of events and attributes in
app component was removed.

Post service is injected now into post-create and post-list components.
An observable subscription is used in post-list to refresh the posts list
whenever the posts in the Post service are updated.